Fix problems compiling some IStream methods on XE
When vilib was converted to work with Delphi 11 some calls to IStream
methods had to be revised because the parameter types had changed
somewhere between Delphi XE and Delphi 11. It is not known when this
happened.

To work around this, a TOleStream was used to wrap the IStream instance
and methods of TOleStream were called instead of the equivalent IStream
methods.

This solution works for both Delphi XE and Delphi 11, presumably because
TOleStream was updated in the VCL when the implementation of IStream
was changed.