Replace use of TVerInfoBinIO with TOleStream
Each parameter reference to TVerInfoBinIO in UVerInfoRec unit was
replaced by a TStream reference.

Wherever TVerInfoBinIO was instantiated in UVerInfoRec, a TOleStream
instance was created instead.

Changes were made to code in UVerInfoRec to account for interface
differences between TVerInfoBinIO and TStream (which were minor).

UVerInfoBinIO was deleted as redundant.